refactor(rslib): collect isolated dts references before codegen
