+++ crates/rspack_javascript_compiler/src/compiler/transform.rs
@@ -5,7 +5,7 @@
  * Author Donny/강동윤
  * Copyright (c)
  */
-use std::{cell::RefCell, fs::File, path::PathBuf, rc::Rc, sync::Arc};
+use std::{cell::RefCell, collections::HashSet, fs::File, path::PathBuf, rc::Rc, sync::Arc};
 
 use anyhow::{Context, bail};
 use indoc::formatdoc;
@@ -27,7 +27,10 @@ use swc_core::{
     errors::Handler,
   },
   ecma::{
-    ast::{EsVersion, Pass, Program},
+    ast::{
+      EsVersion, ExportAll, ImportDecl, NamedExport, Pass, Program, TsExternalModuleRef,
+      TsImportType,
+    },
     codegen::{
       self, Emitter, Node,
       text_writer::{self, WriteJs},
@@ -40,7 +43,7 @@ use swc_core::{
       helpers::{self, Helpers},
       resolver,
     },
-    visit::VisitMutWith,
+    visit::{Visit, VisitMutWith, VisitWith},
   },
 };
 use swc_error_reporters::handler::try_with_handler;
@@ -118,6 +121,11 @@ impl JavaScriptCompiler {
         }
       };
 
+      // Collect declaration references from the FastDts output AST before codegen.
+      // Rslib uses these references to complete type-only declaration outputs without
+      // parsing the generated .d.ts string again.
+      let references = collect_isolated_dts_references(&program);
+
       let code = {
         let mut buf = Vec::new();
         {
@@ -139,7 +147,11 @@ impl JavaScriptCompiler {
         unsafe { String::from_utf8_unchecked(buf) }
       };
 
-      Ok(IsolatedDtsTransformOutput { code, diagnostics })
+      Ok(IsolatedDtsTransformOutput {
+        code,
+        references,
+        diagnostics,
+      })
     })
   }
 
@@ -188,6 +200,62 @@ impl JavaScriptCompiler {
 
       self.emit_isolated_dts(&program, filename, unresolved_mark, target, &comments)
     })
+  }
+}
+
+fn collect_isolated_dts_references(program: &Program) -> Vec<String> {
+  let mut collector = DtsReferenceCollector::default();
+  program.visit_with(&mut collector);
+  collector.references
+}
+
+#[derive(Default)]
+struct DtsReferenceCollector {
+  references: Vec<String>,
+  seen: HashSet<String>,
+}
+
+impl DtsReferenceCollector {
+  fn push(&mut self, value: String) {
+    if self.seen.insert(value.clone()) {
+      self.references.push(value);
+    }
+  }
+}
+
+impl Visit for DtsReferenceCollector {
+  fn visit_import_decl(&mut self, node: &ImportDecl) {
+    // Matches import declarations, for example:
+    //   import type { Foo } from "./foo";
+    //   import "./foo";
+    self.push(node.src.value.to_string_lossy().to_string());
+  }
+
+  fn visit_export_all(&mut self, node: &ExportAll) {
+    // Matches export-all declarations, for example:
+    //   export * from "./foo";
+    self.push(node.src.value.to_string_lossy().to_string());
+  }
+
+  fn visit_named_export(&mut self, node: &NamedExport) {
+    // Matches named re-exports, for example:
+    //   export type { Foo } from "./foo";
+    if let Some(src) = &node.src {
+      self.push(src.value.to_string_lossy().to_string());
+    }
+  }
+
+  fn visit_ts_import_type(&mut self, node: &TsImportType) {
+    // Matches inline import types, for example:
+    //   export type Foo = import("./foo").Foo;
+    self.push(node.arg.value.to_string_lossy().to_string());
+    node.visit_children_with(self);
+  }
+
+  fn visit_ts_external_module_ref(&mut self, node: &TsExternalModuleRef) {
+    // Matches TypeScript import-equals declarations, for example:
+    //   import Foo = require("./foo");
+    self.push(node.expr.value.to_string_lossy().to_string());
   }
 }
